Improve recorder sorting to minimize risk of out-of-order printouts
There will always be out-of-order record dumps in case CPUs contend
for the same recorder. However, an optimization in the previous
sorting algorithm could lead the dump to delay the "most recent" entry
for several lines if it was in a different recorder.

Also added a thread ID in the recorder_test, and a "Multithreading
considerations" with an explanation of the out-of-order timestamps and
orders.