Do not dump string in post-mortemp dumps
For safety reasons, when we have a crash, we do not print strings,
unless you use the special format %!s or %!S, where you guarantee
that the string will always be safe (e.g. text literal).

Strings will be shown during tracing, since they are assumed to be
vaild at the time they are recorded.