Make sure we have both file/line and function information recorded
The __func__ is not a macro, since its value can only been known during
compilation and not preprocessing. So the "where" field is repurposed
to store __func__, whereas file and line info are tagged in front of
the message. A new tweak, recorder_function, controls the output of that
field.

To compensate, the formatting routine has to skip file/line information.
This is controlled by the previous recorder_location tweak.