Print strings safely if dumping a traced recorder
The code was relying incorrectly on rec->trace to decide if the
entry was printed as a result of a dump or of a trace.
However, it is possible to dump a traced entry, so the test
was incorrect. A variable recorder_dumping was added specifically
to indicate when it's unsafe to print a string.