Add command-line options to compute min, max, average and timing
Using -t toggles the display of timing information (with right-hand scale)
Using -m toggles the display of min and max values
Using -a toggles the display of running average value
Using -n toggles the display of the base value, in case you only want timing

The -r option sets the averaging ratio for min, max and average computations.
It is used to make it possible to identify multiple interesting min/max
in a single screen.

Example:

- Run the test with

  RECORDER_TRACES='SpeedTest=tid,value,mod,delay' ./recorder_test_test 1 1000

- Run the scope with:

  recorder_scope -r 99.9 -s sleep_time=0:0:100000 -s sleep_time_delta=0:0:100000 -t mod -m -a -t delay