scope: Add ability to record snapshots of the graph and CSV values
Hitting the 'i' key while the recorder scope is running will record the
current graph as a PNG image file. The current graph is the one that
was clicked last.

Hitting the 'c' key will record the values for the current graph in a
CSV file that can then be imported for example in spreadsheet
applications.

Hitting the space key will record both the CSV values and the image
file.