Clarify the license
The old carried-over headers still had remnants of GPL in them.
Convert to LGPLv3 for what goes in the library, GPLv3 for the rest.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>
Suggested-by: Frediano Ziglio <fziglio@redhat.com>

NOTE: This may not be the final change. If this is to be used
in LGPLv2.1 projects like SPICE, the current setup may not be
correct, as indicated the license compatibility matrix at
https://www.gnu.org/licenses/gpl-faq.html#AllCompatibility.

The intent of the authors is to make the code license-compatible,
and further licensing changes may be required to that effect.