Add comment in the _activate functions
Make it clear that the recorder_ring_compare_exchange macro updates
the head in place with the new value, because it uses
the __atomic_compare_exchange_n primitive, which does, as documented in
https://gcc.gnu.org/onlinedocs/gcc/_005f_005fatomic-Builtins.html

This compares the contents of *ptr with the contents of *expected.
If equal, the operation is a read-modify-write operation that writes
desired into *ptr.
If they are not equal, the operation is a read and the current contents
of *ptr are written into *expected.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>
Suggested-by: Frediano Ziglio <fziglio@redhat.com>