Explicitly update PREFIX.dll when passed --libdir=
While I was not able to reproduce in mock builds, some builds happening in Koji
fail because they install the shared libraries under /usr/lib instead of
/usr/lib64. The only explanation I have for that is that the make-it-quick
configuration on these systems is bogus and points to /usr/lib for PREFIX.dll.

Forcing PREFIX.dll to be updated as well is probably a good idea anyway,
since the configure output should not depend on system-wide (mis)configuration.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>