Add RECORDER_STANDALONE_PRINTF printf emulation
On embedded systems, there is often printf code that is disabled for release
builds. Using RECORDER_STANDALONE_PRINTF converts these to records going to a
recorder called "printf". This shoudl be seen as a stopgap measure, since one of
the benefits of the recorder is that you can sort things by topic.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>